+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Added
 
 - `ui/Touchable` support for drag gesture
+- `ui/Marquee` component
 
 ## [2.0.0-alpha.4] - 2018-02-13
 
@@ -55,7 +56,19 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - `ui/Holdable` and `ui/Pressable` which were replaced by `ui/Touchable`
 
-## [1.13.3] - 2017-01-16
+## [1.15.0] - 2018-02-28
+
+### Fixed
+
+- Internal method used by many components that sometimes prevented re-renders when they were needed
+
+## [1.14.0] - 2018-02-23
+
+### Deprecated
+
+- `ui/Holdable` and `ui/Pressable`, to be replaced by `ui/Touchable` in 2.0.0
+
+## [1.13.3] - 2018-01-16
 
 No significant changes.